compilation warning about printf type mismatch
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@28343 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/http.cpp
@@ -219,7 +219,7 @@ bool wxHTTP::BuildRequest(const wxString& path, wxHTTP_Req req)
   case wxHTTP_POST:
     request = wxT("POST");
     if ( GetHeader( wxT("Content-Length") ).IsNull() )
-      SetHeader( wxT("Content-Length"), wxString::Format( wxT("%ld"), m_post_buf.Len() ) );
+      SetHeader( wxT("Content-Length"), wxString::Format( wxT("%lu"), (unsigned long)m_post_buf.Len() ) );
     break;
   default:
     return FALSE;
